# Speedup AttributeSet#deep_dup